Rename to withI18n and add default module params
Also change to a factory pattern to enable passing params in a compose
chain
